model_llm_service/requirements.txt
flask==3.0.2
python-dotenv==1.0.1
google-generativeai==0.5.4